requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
boto3==1.34.140
botocore==1.34.140
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
Django==5.0.6
django-cors-headers==4.4.0
django-storages==1.14.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
idna==3.8
jiter==0.5.0
jmespath==1.0.1
oauthlib==3.2.2
openai==1.43.0
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pyparsing==3.1.4
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.2
setuptools==72.1.0
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.5
uritemplate==4.1.1
urllib3==2.2.2
wheel==0.43.0
whitenoise==6.7.0
